:wrench: Add global platform defines for the renderer process
